// src/api/generated/models/JwtSignUpRequestDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type JwtSignUpRequestDTO = {
    login?: string;
    password?: string;
    name?: string;
    lastName?: string;
    patronymic?: string;
    phoneNumber?: string;
    organizationName?: string;
    organizationPhoneNumber?: string;
};